Support alpha channel in ImageSharp screenshots
As a size optimization, in the previous commit we set the default pixel
format to RGB24 instead of RGBA32, wrongly assuming that Windows
screenshots do not contain an alpha channel. In fact, they do and not
storing the alpha channel leads to colored artifacts in the resulting
screenshot where semi-transparent pixels were present.